Moved all of the UI to a new file
Added all the UI classes to a new file. This was done to clean up the init file
